Navbar feature (#99)
undefined
